Introduce non static NotificationDataConverter, obsolete old Notifications
