+++ test/Infrastructure/LeanCode.Firebase.FCM.Tests/NotificationConversionTests.cs
@@ -6,12 +6,18 @@ namespace LeanCode.Firebase.FCM.Tests;
 
 public class NotificationConversionTests
 {
-    public NotificationConversionTests() { }
+    private readonly NotificationDataConverter converter;
+
+    public NotificationConversionTests()
+    {
+        converter = new();
+        converter.FormatUsing<DateTimeOffset>(d => d.ToString("O", CultureInfo.InvariantCulture));
+    }
 
     [Fact]
     public void Converts_int_enum_correctly()
     {
-        var data = Notifications.ToNotificationData(new { Field = IntEnum.Second });
+        var data = converter.ToNotificationData(new { Field = IntEnum.Second });
 
         data.Should().ContainKey("Field").WhoseValue.Should().Be("1");
     }
@@ -19,7 +25,7 @@ public class NotificationConversionTests
     [Fact]
     public void Converts_byte_enum_correctly()
     {
-        var data = Notifications.ToNotificationData(new { Field = ByteEnum.Second });
+        var data = converter.ToNotificationData(new { Field = ByteEnum.Second });
 
         data.Should().ContainKey("Field").WhoseValue.Should().Be("1");
     }
@@ -27,10 +33,8 @@ public class NotificationConversionTests
     [Fact]
     public void Applies_custom_value_formatters()
     {
-        Notifications.FormatUsing<DateTimeOffset>(d => d.ToString("O", CultureInfo.InvariantCulture));
-
         var date = new DateTimeOffset(2024, 9, 25, 13, 56, 48, TimeSpan.FromHours(2));
-        var data = Notifications.ToNotificationData(new { Date = date });
+        var data = converter.ToNotificationData(new { Date = date });
 
         data.Should().ContainKey("Date").WhoseValue.Should().Be("2024-09-25T13:56:48.0000000+02:00");
     }
